Added functionality for checkmate and draw verification
Draws were determined according to the rules determined by https://www.chess.com/terms/draw-chess#threefold-repetition
Checkmates just verify if the king is in check and if no legal moves are possible by calling the isStalemate method in the return.
Pressing 'D' on the keyboard now allows to ask for a draw.
Pressing 'R' twice in a row allows the player to resign
Pressing 'X' allows the user to refuse a draw or cancel a resignation request.